fix: typing eof after a decode funciton has been used resulting in infinite loop
The fix is to move the command initializer inside the loop. The problem
was that the residue from the MMU function was still lingering inside stdin,
and this was causing the infinite loop. Now, the loop always resets the
command variable to a null pointer, so the inifinite loop is gone.

Signed-off-by: Soobin Rho <soobinrho@gmail.com>